# backend/API/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          The HTTP status code of the response.
          example: 400
        type: integer
      error:
        description: |-
          The standard HTTP status text for the code.
          example: Bad Request
        type: string
      message:
        description: |-
          A user-friendly message describing the specific issue.
          example: Invalid request body or validation failed.
        type: string
    type: object
...
  models.GroupWithOwner:
    properties:
      created_at:
        description: |-
          The timestamp when the group was created
          example: 2025-12-29T18:00:00Z
        type: string
      description:
        description: |-
          The description of the group
          example: A group for sharing photography tips and showcasing work
        type: string
      id:
        description: |-
          The unique identifier for the group (UUID)
          example: g1a2b3c4-d5e6-f7g8-h9i0-j1k2l3m4n5o6
        type: string
      member_count:
        description: |-
          The number of members in the group
          example: 42
        type: integer
      owner_id:
        description: |-
          The ID of the owner
          example: a1b2c3d4-e5f6-g7h8-i9j0-k1l2m3n4o5p6
        type: string
      owner_nickname:
        description: |-
          The display name of the owner
          example: photo_master
        type: string
      title:
        description: |-
          The title of the group
          example: Photography Enthusiasts
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the social-network application.
  title: Social-network Application API
  version: "1.0"
paths:
...
  /api/v1/groups:
    get:
      description: Retrieves all groups with owner details and member counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GroupWithOwner'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Browse all groups
      tags:
      - Groups
  /api/v1/groups/{id}:
    delete:
      description: Deletes a group (owner only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete group
      tags:
      - Groups
    get:
      description: Retrieves group details with owner information and member count
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GroupWithOwner'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get group details
      tags:
      - Groups
    put:
      consumes:
      - application/json
      description: Updates group details (owner only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated group details
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update group
      tags:
      - Groups
...
  /api/v1/groups/create:
    post:
      consumes:
      - application/json
      description: Creates a new group with the authenticated user as owner and optionally
        invites members
      parameters:
      - description: Group details (title, description, invitees)
        in: body
        name: group
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GroupWithOwner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create a group
      tags:
      - Groups
...
  /api/v1/groups/my-groups:
    get:
      description: Retrieves all groups where the authenticated user is a member
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GroupWithOwner'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get user's groups
      tags:
      - Groups
...
swagger: "2.0"